rename registry -> libraryRegistry
