+++ js/integration/diagram.js
@@ -1,4 +1,4 @@
 const Diagram = require('devexpress-diagram');
-const addLibrary = require('../core/registry').addLibrary;
+const addLibrary = require('../core/library_registry').addLibrary;
 
 Diagram && addLibrary('diagram', Diagram);
