+++ js/integration/gantt.js
@@ -1,4 +1,4 @@
 const Gantt = require('devexpress-gantt');
-const addLibrary = require('../core/registry').addLibrary;
+const addLibrary = require('../core/library_registry').addLibrary;
 
 Gantt && addLibrary('gantt', Gantt);
