+++ js/integration/jszip.js
@@ -1,4 +1,4 @@
 import JSZip from 'jszip';
-import { addLibrary } from '../core/registry';
+import { addLibrary } from '../core/library_registry';
 
 JSZip && addLibrary('JSZip', JSZip);
