+++ js/integration/quill.js
@@ -1,4 +1,4 @@
 import Quill from 'quill';
-import { addLibrary } from '../core/registry';
+import { addLibrary } from '../core/library_registry';
 
 Quill && addLibrary('quill', Quill);
