+++ js/ui/html_editor/formats/align.js
@@ -1,5 +1,5 @@
 
-import { getLibrary } from '../../../core/registry';
+import { getLibrary } from '../../../core/library_registry';
 
 const quill = getLibrary('quill');
 const AlignStyle = quill.import('attributors/style/align');
