+++ js/ui/html_editor/formats/mention.js
@@ -1,5 +1,5 @@
 
-import { getLibrary } from '../../../core/registry';
+import { getLibrary } from '../../../core/library_registry';
 
 const quill = getLibrary('quill');
 const Embed = quill.import('blots/embed');
