+++ js/ui/html_editor/formats/size.js
@@ -1,4 +1,4 @@
-import { getLibrary } from '../../../core/registry';
+import { getLibrary } from '../../../core/library_registry';
 
 const quill = getLibrary('quill');
 const SizeStyle = quill.import('attributors/style/size');
