+++ js/ui/html_editor/quill_registrator.js
@@ -1,4 +1,4 @@
-import { getLibrary } from '../../core/registry';
+import { getLibrary } from '../../core/library_registry';
 
 
 class QuillRegistrator {
