+++ js/ui/html_editor/themes/base.js
@@ -1,4 +1,4 @@
-import { getLibrary } from '../../../core/registry';
+import { getLibrary } from '../../../core/library_registry';
 
 const Theme = getLibrary('quill').import('core/theme');
