+++ testing/tests/DevExpress.ui.widgets/diagramParts/importDiagram.tests.js
@@ -1,5 +1,5 @@
 
-import { getLibrary } from 'core/registry';
+import { getLibrary } from 'core/library_registry';
 import 'integration/diagram';
 
 QUnit.module('Import devexpress-diagram', function() {
